requirements.txt
flask
pymysql
gunicorn
python-dotenv
cryptography
